Extract scan diagnostic path helpers partial
